+++ src/layouts/InterfaceLayout/components/InterfaceAds/InterfaceAds.vue
@@ -23,14 +23,14 @@
       <b-carousel
         v-model="slide"
         :interval="adInterval"
-        controls
         indicators
+        controls
         background="#fff"
         @sliding-start="onSlideStart"
         @sliding-end="onSlideEnd"
       >
         <b-carousel-slide>
-          <mew-connect-ad />
+          <mew-wallet-ad />
         </b-carousel-slide>
 
         <b-carousel-slide>
@@ -47,7 +47,7 @@ import ads from './adComponents';
 export default {
   components: {
     'buy-eth-ad': ads.buyEthAd,
-    'mew-connect-ad': ads.mewConnectAd
+    'mew-wallet-ad': ads.mewWalletAd
   },
   props: {},
   data() {