+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### Devop
+
+- Update greenkeeper packages [#2131](https://github.com/MyEtherWallet/MyEtherWallet/pull/2131)
+
 ### Release v5.4.3
 
 ### Devop