+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### Bug
+
+- Fix button positions [#2125](https://github.com/MyEtherWallet/MyEtherWallet/pull/2125)
+
 ### Devop
 
 - Update BitBox Logo [#2120](https://github.com/MyEtherWallet/MyEtherWallet/pull/2120)